pyproject.toml
[tool.poetry]
name = "psitexamcellbackend"
version = "0.1.0"
description = ""
authors = ["Sanat <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
djangorestframework = "^3.14.0"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.7"
bcrypt = "^4.0.1"
pyjwt = "^2.8.0"
pandas = "^2.1.0"
openpyxl = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
